Moving Contentful tests to dedicated file.
